refactor: move code from ScenarioTestGroup factory method to constructor
